Moved dialog and fragment classes to ui package
