+++ app/src/main/res/layout/activity_main.xml
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/listFragment"
-        android:name="io.github.tstewart.todayi.fragments.AccomplishmentListFragment"
+        android:name="io.github.tstewart.todayi.ui.fragment.AccomplishmentListFragment"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_weight="1" />
